Split out `TObject`.
