+++ local-modules/typecheck/Typecheck.js
@@ -2,8 +2,6 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { ObjectUtil } from 'util-common';
-
 import TypeError from './TypeError';
 
 /**
@@ -121,56 +119,6 @@ export default class Typecheck {
         || (value < minInc)
         || (value > maxInc)) {
       return TypeError.badValue(value, 'int', `${minInc} <= value <= ${maxInc}`);
-    }
-
-    return value;
-  }
-
-  /**
-   * Checks a value of type `object`.
-   *
-   * @param {*} value Value to check.
-   * @returns {object} `value`.
-   */
-  static object(value) {
-    if (typeof value !== 'object') {
-      return TypeError.badValue(value, 'object');
-    }
-
-    return value;
-  }
-
-  /**
-   * Checks a value of type `object`, which must have exactly the indicated set
-   * of keys as "own" properties.
-   *
-   * @param {*} value Value to check.
-   * @param {Array<string>} keys Keys that must be present in `value`.
-   * @returns {object} `value`.
-   */
-  static objectWithExactKeys(value, keys) {
-    if (typeof value !== 'object') {
-      return TypeError.badValue(value, 'object');
-    }
-
-    // Make a copy, check for and delete allowed keys, and see if anything's
-    // left.
-
-    const copy = Object.assign({}, value);
-    for (const k of keys) {
-      if (!ObjectUtil.hasOwnProperty(copy, k)) {
-        return TypeError.badValue(value, 'object', `Missing key \`${k}\``);
-      }
-      delete copy[k];
-    }
-
-    const remainingKeys = Object.keys(copy);
-    if (remainingKeys.length !== 0) {
-      let msg = 'Extra keys';
-      for (const k of remainingKeys) {
-        msg += ` \`${k}\``;
-      }
-      return TypeError.badValue(value, 'object', msg);
     }
 
     return value;
