+++ local-modules/typecheck/main.js
@@ -2,8 +2,9 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
+import TObject from './TObject';
 import TString from './TString';
 import TypeError from './TypeError';
 import Typecheck from './Typecheck';
 
-export { TString, TypeError, Typecheck };
+export { TObject, TString, TypeError, Typecheck };
